+++ src/imamll.h
@@ -30,10 +30,11 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
+#include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
-#include <stdint.h>
-    
+
 #define IMAM_LL_VERSION 15
     
 enum imamLL_errors {
@@ -66,7 +67,7 @@ struct imamLL {
 /* structure of an element */
 struct imamLL_element {
     void *data;                         /* pointer to the data */
-    uint64_t size;                      /* size of data in bytes */
+    size_t size;                      /* size of data in bytes */
     struct imamLL_element *prev;        /* pointer to the next element */
     struct imamLL_element *next;        /* pointer to the first element */
 };
@@ -91,7 +92,7 @@ extern void imamLL_list_error (struct imamLL *list, char *error_message);
 
 /* allocates memory for a new element and returns the pointer of allocated element */
 /* Return Value: upon successful returns pointer to newly allocated element, otherwise returns NULL and sets appropriate error no to error variable of a list*/
-extern struct imamLL_element *imamLL_element_add (struct imamLL *list, uint64_t element_size, uint8_t position);
+extern struct imamLL_element *imamLL_element_add (struct imamLL *list, size_t element_size, uint8_t position);
 
 /* frees and removes the first matching element *element from the list */
 /* Return Value: upon successful returns 1, otherwise returns 0 and -1 and sets appropriate error no to error variable of a list*/
@@ -99,7 +100,7 @@ extern int imamLL_element_remove (struct imamLL *list, struct imamLL_element *el
 
 /* returns the pointer of a matching element */
 /* Return Value: upon successful returns pointer to first found element in a list, otherwise returns NULL and sets appropriate error no to error variable of a list*/
-extern struct imamLL_element *imamLL_element_get (struct imamLL *list, void *element_data, uint64_t data_size);
+extern struct imamLL_element *imamLL_element_get (struct imamLL *list, void *element_data, size_t data_size);
 
 /* returns the pointer of the next element in the list */
 /* Return Value: upon successful returns pointer to the next element in a list, otherwise returns NULL*/
